+++ Libplanet.Net.Tests/Consensus/ContextNonProposerTest.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Security.Cryptography;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -367,7 +366,6 @@ namespace Libplanet.Net.Tests.Consensus
             var unsignedInvalidTx = new UnsignedTx(
                 new TxInvoice(
                     blockChain.Genesis.Hash,
-                    ImmutableHashSet<Address>.Empty,
                     DateTimeOffset.UtcNow,
                     new TxActionList((IValue)List.Empty.Add(new Text("Foo")))), // Invalid action
                 new TxSigningMetadata(txSigner.PublicKey, 0));