+++ Libplanet.Tests/Fixtures/IntegerSet.cs
@@ -57,16 +57,12 @@ namespace Libplanet.Tests.Fixtures
                 .Select(pair => new { State = (BigInteger)pair.State, pair.Key })
                 .Select(pair => new { Action = Arithmetic.Add(pair.State), pair.Key })
                 .Select(pair =>
-                    Transaction.Create(
-                        0,
-                        pair.Key,
-                        null,
-                        new[] { pair.Action }.ToPlainValues(),
-                        null,
-                        null,
-                        ImmutableHashSet<Address>.Empty.Add(pair.Key.Address)
-                    )
-                )
+                    new Transaction(
+                        new UnsignedTx(
+                            new TxInvoice(
+                                actions: new TxActionList(new[] { pair.Action.PlainValue })),
+                            new TxSigningMetadata(pair.Key.PublicKey, 0)),
+                        pair.Key))
                 .OrderBy(tx => tx.Id)
                 .ToImmutableArray();
             Miner = new PrivateKey();