+++ Libplanet.Tests/Tx/TransactionTest.cs
@@ -137,7 +137,6 @@ namespace Libplanet.Tests.Tx
                 }.Select(x => x.PlainValue),
                 null,
                 null,
-                ImmutableHashSet<Address>.Empty,
                 timestamp
             );
 
@@ -179,25 +178,8 @@ namespace Libplanet.Tests.Tx
                 0,
                 _fx.PrivateKey1,
                 null,
-                Array.Empty<DumbAction>().Select(x => x.PlainValue)
-            );
+                Array.Empty<DumbAction>().Select(x => x.PlainValue));
             Assert.Empty(emptyTx.UpdatedAddresses);
-
-            Address updatedAddr = new PrivateKey().Address;
-            var txWithAddr = Transaction.Create(
-                0,
-                _fx.PrivateKey1,
-                null,
-                _fx.TxWithActions.Actions,
-                null,
-                null,
-                new[] { updatedAddr }.ToImmutableHashSet()
-            );
-
-            Assert.Equal(
-                new[] { updatedAddr }.ToHashSet(),
-                txWithAddr.UpdatedAddresses.ToHashSet()
-            );
         }
 
         [Fact]
@@ -210,9 +192,7 @@ namespace Libplanet.Tests.Tx
                 null,
                 Array.Empty<DumbAction>().Select(x => x.PlainValue),
                 null,
-                null,
-                ImmutableHashSet<Address>.Empty
-            );
+                null);
             DateTimeOffset rightAfter = DateTimeOffset.UtcNow;
 
             Assert.InRange(tx.Timestamp, rightBefore, rightAfter);
@@ -230,7 +210,6 @@ namespace Libplanet.Tests.Tx
                     Array.Empty<DumbAction>().Select(x => x.PlainValue),
                     null,
                     null,
-                    ImmutableHashSet<Address>.Empty,
                     DateTimeOffset.UtcNow
                 )
             );
@@ -344,8 +323,9 @@ namespace Libplanet.Tests.Tx
                 genesisHash,
                 updatedAddresses,
                 timestamp,
-                actions
-            );
+                actions,
+                null,
+                null);
             var privateKey =
                 new PrivateKey("51fb8c2eb261ed761429c297dd1f8952c8ce327d2ec2ec5bcc7728e3362627c2");
             PublicKey publicKey = privateKey.PublicKey;
@@ -372,10 +352,11 @@ namespace Libplanet.Tests.Tx
             {
                 var diffInvoice = new TxInvoice(
                     i == 0 ? (BlockHash?)null : invoice.GenesisHash,
-                    i == 1 ? null : invoice.UpdatedAddresses,
-                    i == 2 ? (DateTimeOffset?)DateTimeOffset.MinValue : invoice.Timestamp,
-                    i == 3 ? null : invoice.Actions
-                );
+                    i == 1 ? AddressSet.Empty : invoice.UpdatedAddresses,
+                    i == 2 ? DateTimeOffset.MinValue : invoice.Timestamp,
+                    i == 3 ? TxActionList.Empty : invoice.Actions,
+                    null,
+                    null);
                 var diffSigningMetadata = new TxSigningMetadata(
                     i == 4 ? wrongKey.PublicKey : signingMetadata.PublicKey,
                     i == 5 ? 456L : signingMetadata.Nonce
@@ -422,8 +403,9 @@ namespace Libplanet.Tests.Tx
                 genesisHash,
                 updatedAddresses,
                 timestamp,
-                actions
-            );
+                actions,
+                null,
+                null);
             var privateKey =
                 new PrivateKey("51fb8c2eb261ed761429c297dd1f8952c8ce327d2ec2ec5bcc7728e3362627c2");
             PublicKey publicKey = privateKey.PublicKey;