+++ Libplanet.Tests/Tx/TxFixture.cs
@@ -80,16 +80,18 @@ namespace Libplanet.Tests.Tx
                     ZoneId = 10,
                 },
             };
-            TxWithActions = Transaction.Create(
-                0,
-                PrivateKey1,
-                genesisHash,
-                actions.ToPlainValues(),
-                updatedAddresses: ImmutableHashSet.Create(
-                    new Address("c2a86014073d662a4a9bfcf9cb54263dfa4f5cbc")
-                ),
-                timestamp: timestamp
-            );
+            TxWithActions = new Transaction(
+                new UnsignedTx(
+                    new TxInvoice(
+                        genesisHash: genesisHash,
+                        updatedAddresses: ImmutableHashSet.Create(
+                            new Address("c2a86014073d662a4a9bfcf9cb54263dfa4f5cbc")),
+                        timestamp: timestamp,
+                        actions: new TxActionList(actions.ToPlainValues()),
+                        maxGasPrice: null,
+                        gasLimit: null),
+                    new TxSigningMetadata(PrivateKey1.PublicKey, 0)),
+                PrivateKey1);
         }
 
         public PrivateKey PrivateKey1 { get; }